pyproject.toml
[project]
name = "backend-student-ksenia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.0",
    "pydantic[email]>=2.0",
    "granian>=2.7.1",
    "jamlib>=2.5.5",
    "punq>=0.7.0",
    "tortoise-orm[asyncpg]>=0.25.3",
    "uvloop>=0.22.1",
    "aerich[toml]>=0.9.2",
    "redis[asyncio]>=5.0.0",
    "celery>=5.6.2",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "fakeredis>=2.34.0",
    "httpx>=0.27",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "requests>=2.32.5",
    "ruff>=0.15.0",
    "asyncpg>=0.30.0",
    "icecream>=2.1.10",
    "coverage>=7.13.4",
    "flower>=2.0.1",
]

[tool.aerich]
tortoise_orm = "src.main.AERICH_CONFIG"
location = "./migrations"
src_folder = "./."

[tool.ruff]
line-length = 80
target-version = "py313"
exclude = ["*__init__.py", "migrations"]

[tool.ruff.lint]
extend-select = ["UP", "D"]
fixable = ["ALL"]
ignore = [
    "UP009",
    "D"
]

[tool.ruff.format]
skip-magic-trailing-comma = false
docstring-code-format = false
docstring-code-line-length = 80

[tool.ruff.lint.isort]
case-sensitive = true
constants = ["__future__"]
detect-same-package = true
known-first-party = ["src"]
lines-after-imports = 2
lines-between-types = 1
order-by-type = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = 3.13
ignore_missing_imports = true
disable_error_code = ["import-untyped"]
